Refactor imports to use modelmeta module for consistency
